Change Character::get_skill_level to return an integer
Now it returns just the plain skill level, not the `SkillLevel` object. Most callers only need this anyway.

Code that required the `SkillLevel` object has already been changed in the previous commit to call the new function `get_skill_level_object`.

This removes unnecessary casts of the result of `get_skill_level`:
- to `int` (which was done to extract the actual level from the `SkillLevel` object),
- to `float`/`double` (which is now done implicitly).

Calls to `SkillLevel::level()` which is now done by the function itself.

Assign the result to a plain `int`, not a `SkillLevel` object/reference.

+++ src/iuse_actor.cpp
@@ -1603,10 +1603,10 @@ long enzlave_actor::use( player &p, item &it, bool t, const tripoint& ) const
         add_msg(m_bad, _("You feel horrible for mutilating and enslaving someone's corpse."));
 
         /** @EFFECT_SURVIVAL decreases moral penalty and duration for enzlavement */
-        int moraleMalus = -50 * (5.0 / (float) p.get_skill_level( skill_survival ));
-        int maxMalus = -250 * (5.0 / (float)p.get_skill_level( skill_survival ));
-        int duration = 300 * (5.0 / (float)p.get_skill_level( skill_survival ));
-        int decayDelay = 30 * (5.0 / (float)p.get_skill_level( skill_survival ));
+        int moraleMalus = -50 * (5.0 / p.get_skill_level( skill_survival ));
+        int maxMalus = -250 * (5.0 / p.get_skill_level( skill_survival ));
+        int duration = 300 * (5.0 / p.get_skill_level( skill_survival ));
+        int decayDelay = 30 * (5.0 / p.get_skill_level( skill_survival ));
 
         if (p.has_trait( trait_PACIFIST )) {
             moraleMalus *= 5;
